Slightly expanded outline for types/concepts talk.
Now I'm happy with the overall structure of the talk. I was a bit confused on how present two related topics (types-as-design and concepts), and I'm still not 100% convinced in the current order/approach, but I've realized that the two concepts are in a sort of cyclic relationship that means there will never be a 100% satisfying linearization of the ideas.